requirements.txt
# Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Frontend
streamlit>=1.52.0
plotly>=6.0.0

# Azure Blob Storage
azure-storage-blob>=12.19.0

# Chemistry
rdkit>=2024.9.0
chembl-webresource-client>=0.10.9
pandas>=2.0.0
numpy>=2.0.0

# Utils
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
